third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGPatternElement-svgdom-patternContentUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternContentUnits' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS patternElement.patternContentUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS successfullyParsed is true

TEST COMPLETE
